refactor(perl-ci-hygiene): split CI policy/process helpers into modules (#0000)
