+++ crates/perl-ci-hygiene/src/process_io.rs
@@ -1,0 +1,87 @@
+use color_eyre::eyre::{Result, WrapErr};
+use std::io::Write;
+use std::path::Path;
+use std::process::{Command, Stdio};
+
+pub(crate) fn command_with_output(
+    repo_root: &Path,
+    command: &str,
+    args: &[&str],
+    env_vars: &[(&str, &str)],
+) -> Result<String> {
+    let mut child = Command::new(command);
+    child.current_dir(repo_root).args(args);
+    for (key, value) in env_vars {
+        child.env(key, value);
+    }
+    child.stdout(Stdio::piped()).stderr(Stdio::piped());
+    let output = child.output().wrap_err_with(|| format!("running {command}"))?;
+    let status = output.status.code().unwrap_or(1);
+    let stdout = String::from_utf8_lossy(&output.stdout).into_owned();
+    if status != 0 {
+        let stderr = String::from_utf8_lossy(&output.stderr).into_owned();
+        return Err(color_eyre::eyre::eyre!(
+            "command '{command}' failed (exit {status}): {stderr}"
+        ));
+    }
+    Ok(stdout)
+}
+
+pub(crate) fn command_with_output_all(
+    repo_root: &Path,
+    command: &str,
+    args: &[&str],
+    env_vars: &[(&str, &str)],
+) -> Result<String> {
+    let mut child = Command::new(command);
+    child.current_dir(repo_root).args(args);
+    for (key, value) in env_vars {
+        child.env(key, value);
+    }
+    child.stdout(Stdio::piped()).stderr(Stdio::piped());
+    let output = child.output().wrap_err_with(|| format!("running {command}"))?;
+    let mut combined = String::from_utf8_lossy(&output.stdout).into_owned();
+    if !output.stderr.is_empty() {
+        combined.push_str(&String::from_utf8_lossy(&output.stderr));
+    }
+    let status = output.status.code().unwrap_or(1);
+    if status != 0 {
+        return Err(color_eyre::eyre::eyre!(
+            "command '{command}' failed (exit {status}): {combined}"
+        ));
+    }
+    Ok(combined)
+}
+
+pub(crate) fn command_with_input_with_status(
+    repo_root: &Path,
+    command: &str,
+    args: &[&str],
+    env_vars: &[(&str, &str)],
+    stdin_payload: &str,
+) -> Result<(i32, String)> {
+    let mut child = Command::new(command);
+    child.current_dir(repo_root).args(args);
+    for (key, value) in env_vars {
+        child.env(key, value);
+    }
+    child.stdin(Stdio::piped()).stdout(Stdio::piped()).stderr(Stdio::piped());
+
+    let mut child = child.spawn().wrap_err_with(|| format!("running {command}"))?;
+    {
+        let mut stdin = child
+            .stdin
+            .take()
+            .ok_or_else(|| color_eyre::eyre::eyre!("failed to open stdin for command {command}"))?;
+        stdin
+            .write_all(stdin_payload.as_bytes())
+            .wrap_err_with(|| format!("writing to stdin for {command}"))?;
+    }
+    let output = child.wait_with_output().wrap_err_with(|| format!("running {command}"))?;
+    let status = output.status.code().unwrap_or(1);
+    let mut combined = String::from_utf8_lossy(&output.stdout).into_owned();
+    if !output.stderr.is_empty() {
+        combined.push_str(&String::from_utf8_lossy(&output.stderr));
+    }
+    Ok((status, combined))
+}
